% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_count.R
\name{check_count}
\alias{check_count}
\title{Check cell counts for categorical variables}
\usage{
check_count(data, limit, col)
}
\arguments{
\item{data}{A data frame or tibble to be used}

\item{limit}{The value for k as an integer}

\item{col}{The variable to be checked in Tidyverse notation}
}
\value{
Data frame containing all variable categories with counts less than a specified limit
}
\description{
Check cell counts for categorical variables
}
\examples{
data <- data.frame(
 A = sample(c(1, 2, 3), 100, replace = TRUE),
 B = sample(c(4, 5), 100, replace = TRUE),
 C = sample(c(6, 7, 8), 100, replace = TRUE)
)
check_count(data, 5, A)
}
