% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_anonymity.R
\name{k_anonymity}
\alias{k_anonymity}
\title{Check multiple categorical variables for k-anonymity}
\usage{
k_anonymity(data, limit, ...)
}
\arguments{
\item{data}{A data frame or tibble to be used}

\item{limit}{The value for k as an integer}

\item{...}{Two or more categorical variables using Tidyverse syntax (see Details)}
}
\value{
Data frame containing all unique combinations whose counts fall under the supplied limit. Returns \code{NULL} if all combinations are above the limit.
}
\description{
Check multiple categorical variables for k-anonymity
}
\details{
The \code{...} parameter supports two or more categorical variables supplied using Tidyverse notation. This means full support for \code{:}, \code{!}, \code{&}, \code{|} and \code{c()}.
}
\examples{
data <- data.frame(
 A = sample(c(1, 2, 3), 100, replace = TRUE),
 B = sample(c(4, 5), 100, replace = TRUE),
 C = sample(c(6, 7, 8), 100, replace = TRUE)
)
# Explicitly specify variables
k_anonymity(data, 5, A, B, C)
# Using quosure slicing
k_anonymity(data, 5, A:C)
# Using negation
k_anonymity(data, 5, !c(A:B))
}
