% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranon-package.R
\docType{package}
\name{ranon-package}
\alias{ranon}
\alias{ranon-package}
\title{ranon: Anonymise Human Data Sets}
\description{
Convenience functions for checking human data set anonymity.
}
\author{
\strong{Maintainer}: Oliver Saal \email{oliver.saal@proton.me} (\href{https://orcid.org/0000-0002-4870-0458}{ORCID})

}
\keyword{internal}
